Refactor files to src folder
